+++ src/main/kotlin/org/kryptonmc/krypton/world/generation/WorldGenerationSettings.kt
@@ -1,9 +1,10 @@
 package org.kryptonmc.krypton.world.generation
 
+import org.kryptonmc.krypton.registry.NamespacedKey
 import org.kryptonmc.krypton.world.dimension.Dimension
 
 data class WorldGenerationSettings(
     val seed: Long,
     val generateStructures: Boolean,
-    val dimensions: List<Dimension>
+    val dimensions: Map<NamespacedKey, Dimension>
 )